+++ tests/Concerns/WithResourceFilteringTest.php
@@ -8,8 +8,6 @@ use XtendPackages\RESTPresenter\Concerns\WithResourceFiltering;
 use XtendPackages\RESTPresenter\Models\User;
 use XtendPackages\RESTPresenter\Resources\Users\Filters\UserEmailVerified;
 
-use function XtendPackages\RESTPresenter\Support\Tests\invokeNonPublicMethod;
-
 beforeEach(function (): void {
     $this->resourceController = new class()
     {
@@ -36,15 +34,15 @@ describe('WithResourceFiltering', function (): void {
         $request->merge(['filters' => ['email_verified_at' => now()]]);
         app()->instance('request', $request);
 
-        $originalQuery = invokeNonPublicMethod($this->resourceController, 'getModelQuery');
-        $newQuery = invokeNonPublicMethod($this->resourceController, 'applyFilters', [clone $originalQuery]);
+        $originalQuery = invade($this->resourceController)->getModelQuery();
+        $newQuery = invade($this->resourceController)->applyFilters(clone $originalQuery);
 
         expect($newQuery->toSql())->toContain('"email_verified_at" is not null')
             ->and($newQuery->toSql())->not()->toBe($originalQuery->toSql());
     });
 
     test('getFilters returns resource filters', function (): void {
-        $filters = invokeNonPublicMethod($this->resourceController, 'getFilters');
+        $filters = invade($this->resourceController)->getFilters();
 
         expect($filters)->toBeArray()
             ->and($filters)->toHaveKey('email_verified_at');